RDCC-2029 refactored functional test cases
